# python-flask/swagger_server/swagger/swagger.yaml
...
swagger: "2.0"
info:
  description: "Ytel API version 3"
  version: "3.12"
  title: "Ytel API V3"
host: "api.ytel.com"
basePath: "/api/v3"
schemes:
- "https"
consumes:
- "application/x-www-form-urlencoded"
produces:
- "application/json"
paths:
...
  /address/viewaddress.json:
    post:
      tags:
      - "Address"
      summary: "View Address"
      description: "View Address Specific address Book by providing the address id"
      operationId: "address_viewaddress"
      consumes:
      - "application/x-www-form-urlencoded"
      produces:
      - "application/json"
      parameters:
      - name: "addressid"
        in: "formData"
        description: "The identifier of the address to be retrieved."
        required: true
        type: "string"
      responses:
        200:
          description: ""
          examples: {}
          schema:
            type: "string"
      security:
      - auth: []
      x-unitTests: []
      x-operation-settings:
        CollectParameters: true
        AllowDynamicQueryParameters: false
        AllowDynamicFormParameters: false
        IsMultiContentStreaming: false
      x-swagger-router-controller: "swagger_server.controllers.address_controller"
  /address/listaddress.json:
    post:
      tags:
      - "Address"
      summary: "List Addresses"
      description: "List All Address "
      operationId: "address_listaddress"
      consumes:
      - "application/x-www-form-urlencoded"
      produces:
      - "application/json"
      parameters:
      - name: "page"
        in: "formData"
        description: "The page count to retrieve from the total results in the collection.\
          \ Page indexing starts at 1."
        required: false
        type: "integer"
        default: 1
        exclusiveMaximum: false
        exclusiveMinimum: false
        format: "int32"
      - name: "pagesize"
        in: "formData"
        description: "How many results to return, default is 10, max is 100, must\
          \ be an integer"
        required: false
        type: "integer"
        default: 10
        exclusiveMaximum: false
        exclusiveMinimum: false
        format: "int32"
      - name: "addressid"
        in: "formData"
        description: "addresses Sid"
        required: false
        type: "string"
      - name: "dateCreated"
        in: "formData"
        description: "date created address."
        required: false
        type: "string"
      responses:
        200:
          description: ""
          examples: {}
          schema:
            type: "string"
      security:
      - auth: []
      x-unitTests: []
      x-operation-settings:
        CollectParameters: true
        AllowDynamicQueryParameters: false
        AllowDynamicFormParameters: false
        IsMultiContentStreaming: false
      x-swagger-router-controller: "swagger_server.controllers.address_controller"
  /address/verifyaddress.json:
    post:
      tags:
      - "Address"
      summary: "Verify Address"
      description: "Validates an address given."
      operationId: "address_verifyaddress"
      consumes:
      - "application/x-www-form-urlencoded"
      produces:
      - "application/json"
      parameters:
      - name: "addressid"
        in: "formData"
        description: "The identifier of the address to be verified."
        required: true
        type: "string"
      responses:
        200:
          description: ""
          examples: {}
          schema:
            type: "string"
      security:
      - auth: []
      x-unitTests: []
      x-operation-settings:
        CollectParameters: true
        AllowDynamicQueryParameters: false
        AllowDynamicFormParameters: false
        IsMultiContentStreaming: false
      x-swagger-router-controller: "swagger_server.controllers.address_controller"
  /address/deleteaddress.json:
    post:
      tags:
      - "Address"
      summary: "Delete Address"
      description: "To delete Address to your address book"
      operationId: "address_deleteaddress"
      consumes:
      - "application/x-www-form-urlencoded"
      produces:
      - "application/json"
      parameters:
      - name: "addressid"
        in: "formData"
        description: "The identifier of the address to be deleted."
        required: true
        type: "string"
      responses:
        200:
          description: ""
          examples: {}
          schema:
            type: "string"
      security:
      - auth: []
      x-unitTests: []
      x-operation-settings:
        CollectParameters: true
        AllowDynamicQueryParameters: false
        AllowDynamicFormParameters: false
        IsMultiContentStreaming: false
      x-swagger-router-controller: "swagger_server.controllers.address_controller"
...
  /letter/viewletter.json:
    post:
      tags:
      - "Letter"
      summary: "View Letter"
      description: "Retrieve a letter object by its LetterSid."
      operationId: "letter_viewletter"
      consumes:
      - "application/x-www-form-urlencoded"
      produces:
      - "application/json"
      parameters:
      - name: "lettersid"
        in: "formData"
        description: "The unique identifier for a letter object."
        required: true
        type: "string"
      responses:
        200:
          description: ""
          examples: {}
          schema:
            type: "string"
      security:
      - auth: []
      x-unitTests: []
      x-operation-settings:
        CollectParameters: true
        AllowDynamicQueryParameters: false
        AllowDynamicFormParameters: false
        IsMultiContentStreaming: false
      x-swagger-router-controller: "swagger_server.controllers.letter_controller"
...
  /letter/create.json:
    post:
      tags:
      - "Letter"
      summary: "Create Letter"
      description: "Create, print, and mail a letter to an address. The letter file\
        \ must be supplied as a PDF or an HTML string."
      operationId: "letter_create"
      consumes:
      - "application/x-www-form-urlencoded"
      produces:
      - "application/json"
      parameters:
      - name: "to"
        in: "formData"
        description: "The AddressId or an object structured when creating an address\
          \ by addresses/Create."
        required: true
        type: "string"
      - name: "from"
        in: "formData"
        description: "The AddressId or an object structured when creating an address\
          \ by addresses/Create."
        required: true
        type: "string"
      - name: "attachbyid"
        in: "formData"
        description: "Set an existing letter by attaching its LetterId."
        required: true
        type: "string"
      - name: "file"
        in: "formData"
        description: "File can be a 8.5\"x11\" PDF uploaded file or URL that links\
          \ to a file."
        required: true
        type: "string"
      - name: "color"
        in: "formData"
        description: "Specify if letter should be printed in color."
        required: true
        type: "string"
      - name: "description"
        in: "formData"
        description: "A description for the letter."
        required: false
        type: "string"
      - name: "extraservice"
        in: "formData"
        description: "Add an extra service to your letter. Options are \"certified\"\
          \ or \"registered\". Certified provides tracking and delivery confirmation\
          \ for domestic destinations and is an additional $5.00. Registered provides\
          \ tracking and confirmation for international addresses and is an additional\
          \ $16.50."
        required: false
        type: "string"
      - name: "doublesided"
        in: "formData"
        description: "Specify if letter should be printed on both sides."
        required: false
        type: "string"
      - name: "template"
        in: "formData"
        description: "Boolean that defaults to true. When set to false, this specifies\
          \ that your letter does not follow the m360 address template. In this case,\
          \ a blank address page will be inserted at the beginning of your file and\
          \ you will be charged for the extra page."
        required: false
        type: "string"
      - name: "htmldata"
        in: "formData"
        description: "A string value that contains HTML markup."
        required: false
        type: "string"
      responses:
        200:
          description: ""
          examples: {}
          schema:
            type: "string"
      security:
      - auth: []
      x-unitTests: []
      x-operation-settings:
        CollectParameters: true
        AllowDynamicQueryParameters: false
        AllowDynamicFormParameters: false
        IsMultiContentStreaming: false
      x-swagger-router-controller: "swagger_server.controllers.letter_controller"
  /letter/listsletter.json:
    post:
      tags:
      - "Letter"
      summary: "List Letters"
      description: "Retrieve a list of letter objects. The letter objects are sorted\
        \ by creation date, with the most recently created letters appearing first."
      operationId: "letter_listsletter"
      consumes:
      - "application/x-www-form-urlencoded"
      produces:
      - "application/json"
      parameters:
      - name: "page"
        in: "formData"
        description: "The page count to retrieve from the total results in the collection.\
          \ Page indexing starts at 1."
        required: false
        type: "integer"
        default: 1
        exclusiveMaximum: false
        exclusiveMinimum: false
        format: "int32"
      - name: "pagesize"
        in: "formData"
        description: "The count of objects to return per page."
        required: false
        type: "integer"
        default: 10
        exclusiveMaximum: false
        exclusiveMinimum: false
        format: "int32"
      - name: "lettersid"
        in: "formData"
        description: "The unique identifier for a letter object."
        required: false
        type: "string"
      - name: "dateCreated"
        in: "formData"
        description: "The date the letter was created."
        required: false
        type: "string"
      responses:
        200:
          description: ""
          examples: {}
          schema:
            type: "string"
      security:
      - auth: []
      x-unitTests: []
      x-operation-settings:
        CollectParameters: true
        AllowDynamicQueryParameters: false
        AllowDynamicFormParameters: false
        IsMultiContentStreaming: false
      x-swagger-router-controller: "swagger_server.controllers.letter_controller"
  /letter/delete.json:
    post:
      tags:
      - "Letter"
      summary: "Delete Letter"
      description: "Remove a letter object by its LetterId."
      operationId: "letter_delete"
      consumes:
      - "application/x-www-form-urlencoded"
      produces:
      - "application/json"
      parameters:
      - name: "lettersid"
        in: "formData"
        description: "The unique identifier for a letter object."
        required: true
        type: "string"
      responses:
        200:
          description: ""
          examples: {}
          schema:
            type: "string"
      security:
      - auth: []
      x-unitTests: []
      x-operation-settings:
        CollectParameters: true
        AllowDynamicQueryParameters: false
        AllowDynamicFormParameters: false
        IsMultiContentStreaming: false
      x-swagger-router-controller: "swagger_server.controllers.letter_controller"
...
  /address/createaddress.json:
    post:
      tags:
      - "Address"
      summary: "Create Address"
      description: "To add an address to your address book, you create a new address\
        \ object. You can retrieve and delete individual addresses as well as get\
        \ a list of addresses. Addresses are identified by a unique random ID."
      operationId: "address_createaddress"
      consumes:
      - "application/x-www-form-urlencoded"
      produces:
      - "application/json"
      parameters:
      - name: "Name"
        in: "formData"
        description: "Name of user"
        required: true
        type: "string"
      - name: "Address"
        in: "formData"
        description: "Address of user."
        required: true
        type: "string"
      - name: "Country"
        in: "formData"
        description: "Must be a 2 letter country short-name code (ISO 3166)"
        required: true
        type: "string"
      - name: "State"
        in: "formData"
        description: "Must be a 2 letter State eg. CA for US. For Some Countries it\
          \ can be greater than 2 letters."
        required: true
        type: "string"
      - name: "City"
        in: "formData"
        description: "City Name."
        required: true
        type: "string"
      - name: "Zip"
        in: "formData"
        description: "Zip code of city."
        required: true
        type: "string"
      - name: "Description"
        in: "formData"
        description: "Description of addresses."
        required: false
        type: "string"
      - name: "email"
        in: "formData"
        description: "Email Id of user."
        required: false
        type: "string"
      - name: "Phone"
        in: "formData"
        description: "Phone number of user."
        required: false
        type: "string"
      responses:
        200:
          description: ""
          examples: {}
          schema:
            type: "string"
      security:
      - auth: []
      x-unitTests: []
      x-operation-settings:
        CollectParameters: true
        AllowDynamicQueryParameters: false
        AllowDynamicFormParameters: false
        IsMultiContentStreaming: false
      x-swagger-router-controller: "swagger_server.controllers.address_controller"
...
securityDefinitions:
  auth:
    type: "basic"
    x-testValue: []
    x-skip-client-authentication: false
...