# rails5/app/controllers/address_controller.rb
=begin
Ytel API V3

Ytel API version 3

OpenAPI spec version: 3.12

Generated by: https://github.com/swagger-api/swagger-codegen.git

=end
class AddressController < ApplicationController

  def address_createaddress
    # Your code here

    render json: {"message" => "yes, it worked"}
  end

  def address_deleteaddress
    # Your code here

    render json: {"message" => "yes, it worked"}
  end

  def address_listaddress
    # Your code here

    render json: {"message" => "yes, it worked"}
  end

  def address_verifyaddress
    # Your code here

    render json: {"message" => "yes, it worked"}
  end

  def address_viewaddress
    # Your code here

    render json: {"message" => "yes, it worked"}
  end
end