# rails5/app/controllers/letter_controller.rb
=begin
Ytel API V3

Ytel API version 3

OpenAPI spec version: 3.12

Generated by: https://github.com/swagger-api/swagger-codegen.git

=end
class LetterController < ApplicationController

  def letter_create
    # Your code here

    render json: {"message" => "yes, it worked"}
  end

  def letter_delete
    # Your code here

    render json: {"message" => "yes, it worked"}
  end

  def letter_listsletter
    # Your code here

    render json: {"message" => "yes, it worked"}
  end

  def letter_viewletter
    # Your code here

    render json: {"message" => "yes, it worked"}
  end
end